requirements.txt
fastapi
uvicorn[standard]
python-dotenv
zyte_api
openai
pydantic
